Move mjd_to_day to util_functions
