+++ src/kbmod/util_functions.py
@@ -48,6 +48,22 @@ def get_matched_obstimes(obs_times, query_times, threshold=0.0007):
     return min_inds
 
 
+def mjd_to_day(mjd):
+    """Takes an mjd and converts it into a day in calendar date format.
+
+    Parameters
+    ----------
+    mjd : `float`
+        mjd format date.
+
+    Returns
+    ----------
+    A `str` with a calendar date, in the format YYYY-MM-DD.
+    e.g., mjd=60000 -> '2023-02-25'
+    """
+    return Time(mjd, format="mjd").strftime("%Y-%m-%d")
+
+
 def load_deccam_layered_image(filename, psf):
     """Load a layered image from the legacy deccam format.
