+++ tests/test_util_functions.py
@@ -7,6 +7,7 @@ from kbmod.search import LayeredImage, PSF
 from kbmod.util_functions import (
     get_matched_obstimes,
     load_deccam_layered_image,
+    mjd_to_day,
 )
 
 
@@ -17,6 +18,11 @@ class test_util_functions(unittest.TestCase):
         matched_inds = get_matched_obstimes(obstimes, query_times, threshold=0.01)
         expected = [-1, 0, 0, -1, 1, 4, 5, 5, 7, 7, -1]
         self.assertTrue(np.array_equal(matched_inds, expected))
+
+    def test_mjd_to_day(self):
+        self.assertEqual(mjd_to_day(60481.04237269), "2024-06-20")
+        self.assertEqual(mjd_to_day(60500), "2024-07-09")
+        self.assertEqual(mjd_to_day(58100.5), "2017-12-13")
 
     def test_load_deccam_layered_image(self):
         base_path = Path(__file__).parent.parent
